Resolve Jest dependency version, move form files to components folder
