+++ package.json
@@ -42,7 +42,6 @@
   },
   "devDependencies": {
     "@babel/preset-react": "^7.12.10",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "jest": "^26.6.3"
+    "babel-plugin-transform-class-properties": "^6.24.1"
   }
 }
